move VehiclesByPosition and groupByPosition to its own model/ file
